+++ fractal/components/assessments/RecommendationDetail/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import RecommendationDetail from './RecommendationDetail.jsx';
+import config from './RecommendationDetail.config';
+
+ReactDOM.render(
+    <div>
+        <RecommendationDetail {...config.context} />
+        <RecommendationDetail {...config.variants[0].context} />
+    </div>,
+    document.getElementsByClassName('base-preview')[0]
+);