+++ fractal/components/assessments/RecommendationTOC/RecommendationTOC.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+    "title": "Assessment Recommendation Table of Contents",
+    "status": "ready",
+    "collated": true,
+    "preview": "@recommendation-toc-preview",
+    "context": {
+        imageTitle: 'Lorem Ipsum',
+        imageUrl: 'https://placeimg.com/320/240/animals',
+        baseUrl: '/somewhere',
+        title: 'IMPROVE YOUR SCORE',
+        tableOfContents: [
+            {to: 'a', text: 'Item 1'},
+            {to: 'b', text: 'Item 2'},
+            {to: 'c', text: 'Item 3'}
+        ]
+    }
+};