+++ fractal/components/links/HashLink/HashLink.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    "title": "Link",
+    "status": "ready",
+    "collated": true,
+    "preview": "@preview",
+    "context": {
+        "to": "http://www.google.com",
+        "label": "Test Link"
+    }
+};